Unify gcm and gcm2 interfaces
